No longer setting a pixel font size on the body.
This better respects the browser default size set by the user. The right way to scale typography globaly for large or small fonts (per project) is a percentage font size on the html tag, in order to take into account the user's setting and influence how rem units are calculated.